requirements.txt
# --- For-Bot-Working --------- #
pyrofork
TgCrypto
pyromod==1.5
python-dotenv
requests
pyshorteners
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp

